Use ratio framebuffer_size/window_size to fix gui widgets in macOS
In macOS, highdpi screens create a scaled framebuffer that is then
shrunken to fit into the logical window size.  This leads to errors in
which the screen position as obtained from the window system differs
from the actual pixel's positions on the OpenGL framebuffer, which
affects the rendering of textures and picking.